+++ src/main/java/com/TwoSeaU/BaData/domain/store/dto/response/StoreResponse.java
@@ -1,6 +1,5 @@
-package com.TwoSeaU.BaData.domain.store.dto;
+package com.TwoSeaU.BaData.domain.store.dto.response;
 
-import com.TwoSeaU.BaData.domain.store.entity.Store;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
